refactor: eliminate intermediate barrel files to simplify and for consistency
