--- libs/angular/src/lib/adapters/index.ts
@@ -1,4 +1,0 @@
-export { ActionWidgetAdapter } from './action-widget-adapter.service';
-export { DisplayWidgetAdapter } from './display-widget-adapter.service';
-export { InputWidgetAdapter } from './input-widget-adapter.service';
-export { LayoutWidgetAdapter } from './layout-widget-adapter.service';
