+++ libs/core/src/index.ts
@@ -36,7 +36,7 @@ export type {
 
 export type { ItemRenderContext, ItemRenderItemData } from './lib/item-renderer';
 
-export { devToolsMiddleware } from './lib/middleware';
+export { devToolsMiddleware } from './lib/middleware/devtools/devtools.middleware';
 
 export type {
   ActionWidgetTemplateData,
