--- libs/core/src/lib/context/index.ts
@@ -1,4 +1,0 @@
-export { FormContext } from './form.context';
-export { WidgetRegistry } from './widget-registry';
-export type { WidgetLoaders } from './widget-registry';
-export type { WithWidget } from './with-widget.type';
