--- libs/core/src/lib/middleware/index.ts
@@ -1,1 +1,0 @@
-export { devToolsMiddleware } from './devtools/devtools.middleware';
