--- libs/gui/components/src/lib/components/index.ts
@@ -1,30 +1,0 @@
-export { weekInfoData } from '../utils/week-info';
-export type { WeekInfo } from '../utils/week-info';
-export { GuiButton } from './button';
-export { GuiCalendar } from './calendar';
-export type { CalendarDay } from './calendar';
-export { GuiCheckbox } from './checkbox';
-export { GuiCurrency } from './currency';
-export { GuiDate } from './date-input';
-export { GuiErrors } from './errors';
-export { GuiLabel } from './label';
-export { GuiList } from './list';
-export { GuiMarkdown } from './markdown';
-export { GuiMarkdownText } from './markdown-text';
-export { GuiNumber } from './number';
-export {
-  createOptionMapper,
-  inferOptionValue,
-  isOptionValue,
-  isProtoOption,
-  updateOptions,
-} from './one-of';
-export { GuiPassword } from './password';
-export { GuiRadiogroup } from './radiogroup';
-export { GuiRangeCalendar } from './range-calendar';
-export type { RangeCalendarDay } from './range-calendar';
-export { GuiRangeDateInput } from './range-date-input';
-export { GuiSelect } from './select';
-export { GuiTextarea } from './textarea';
-export { GuiTextinput } from './textinput';
-export { GuiToggle } from './toggle';
