+++ libs/gui/components/src/lib/components/select.ts
@@ -1,15 +1,15 @@
-import { html, LitElement, nothing } from 'lit';
-import { repeat } from 'lit-html/directives/repeat.js';
-import { customElement, property } from 'lit/decorators.js';
-import { classMap } from 'lit/directives/class-map.js';
-import { GUIAriaController } from '../controllers';
-import { addErrors, addIcon, addLabel, type ControlTemplateData } from '../utils/templates';
 import {
   type OneOfProps,
   type Option,
   type OptionValue,
   type SelectProps,
 } from '@golemui/gui-shared';
+import { html, LitElement, nothing } from 'lit';
+import { repeat } from 'lit-html/directives/repeat.js';
+import { customElement, property } from 'lit/decorators.js';
+import { classMap } from 'lit/directives/class-map.js';
+import { GUIAriaController } from '../controllers/aria.controller';
+import { addErrors, addIcon, addLabel, type ControlTemplateData } from '../utils/templates';
 import { inferOptionValue, updateOptions } from './one-of';
 
 @customElement('gui-select')
