--- libs/gui/components/src/lib/controllers/index.ts
@@ -1,1 +1,0 @@
-export { GUIAriaController } from './aria.controller';
