--- libs/gui/shared/src/lib/utils/index.ts
@@ -1,3 +1,0 @@
-export { isOption } from './one-of';
-export { getItemKey } from './repeater';
-export { resolveChunkRefs } from './schema';
