--- libs/react/src/lib/utils/index.ts
@@ -1,2 +1,0 @@
-export { cn } from './cn';
-export type { Cn, CnArray, CnRecord, CnValue } from './cn';
